# docs/openapi.yaml
openapi: 3.0.0
info:
  title: 'AZE FARM API'
  version: 1.0.0
  description: 'API pour la gestion des produits agricoles'
servers:
  -
    url: 'http://localhost:3000/api'
    description: 'Serveur de développement'
paths:
...
  '/api/oauth/{provider}':
    get:
      summary: 'Initier l''authentification OAuth 2.0'
      tags:
        - OAuth
      description: "Redirige vers le fournisseur OAuth pour l'authentification.\n\n**Fournisseurs supportés:**\n- google\n- github\n- facebook\n- instagram\n- twitter\n- linkedin\n"
      parameters:
        -
          in: path
          name: provider
          required: true
          schema:
            type: string
            enum:
              - google
              - github
              - facebook
              - instagram
              - twitter
              - linkedin
          description: 'Fournisseur OAuth'
        -
          in: query
          name: redirect_uri
          schema:
            type: string
          description: 'URI de redirection après authentification (optionnel)'
      responses:
        '302':
          description: 'Redirection vers le fournisseur OAuth'
        '400':
          description: 'Fournisseur non supporté'
        '500':
          description: 'Erreur de configuration OAuth'
  '/api/oauth/{provider}/callback':
    get:
      summary: 'Callback OAuth 2.0 (ne pas appeler directement)'
      tags:
        - OAuth
      description: "Point de terminaison de callback appelé par le fournisseur OAuth après l'autorisation.\n\n**Cette route ne doit pas être appelée directement par les clients.**\nElle est appelée automatiquement par le fournisseur OAuth.\n"
      parameters:
        -
          in: path
          name: provider
          required: true
          schema:
            type: string
            enum:
              - google
              - github
              - facebook
              - instagram
              - twitter
              - linkedin
          description: 'Fournisseur OAuth'
        -
          in: query
          name: code
          schema:
            type: string
          description: 'Code d''autorisation OAuth'
        -
          in: query
          name: state
          schema:
            type: string
          description: 'État OAuth pour la sécurité CSRF'
        -
          in: query
          name: error
          schema:
            type: string
          description: 'Erreur OAuth (si applicable)'
      responses:
        '302':
          description: 'Redirection vers l''application frontend avec tokens'
          headers:
            Location:
              schema:
                type: string
              description: 'URL de redirection avec tokens en paramètres'
        '400':
          description: 'Erreur d''authentification OAuth'
        '500':
          description: 'Erreur serveur lors du traitement OAuth'
  /api/oauth/accounts:
    get:
      summary: 'Récupérer les comptes OAuth liés'
      tags:
        - OAuth
      security:
        -
          bearerAuth: []
      description: 'Retourne la liste des fournisseurs OAuth liés au compte utilisateur'
      responses:
        '200':
          description: 'Comptes OAuth récupérés avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Comptes OAuth récupérés'}
                  data: {type: object, properties: {accounts: {type: array, items: {type: object, properties: {provider: {type: string, enum: [google, github, facebook, instagram, twitter, linkedin], example: google}, providerId: {type: string, description: 'ID utilisateur chez le fournisseur'}, email: {type: string, description: 'Email du compte OAuth'}, displayName: {type: string, description: 'Nom d''affichage'}, linkedAt: {type: string, format: date-time, description: 'Date de liaison du compte'}}}}}}
        '401':
          description: 'Non authentifié'
        '500':
          description: 'Erreur serveur'
  '/api/oauth/{provider}/unlink':
    delete:
      summary: 'Délier un compte OAuth'
      tags:
        - OAuth
      security:
        -
          bearerAuth: []
      description: 'Supprime la liaison entre le compte utilisateur et un fournisseur OAuth'
      parameters:
        -
          in: path
          name: provider
          required: true
          schema:
            type: string
            enum:
              - google
              - github
              - facebook
              - instagram
              - twitter
              - linkedin
          description: 'Fournisseur OAuth à délier'
      responses:
        '200':
          description: 'Compte OAuth délié avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Compte OAuth délié avec succès'}
        '400':
          description: 'Fournisseur invalide ou compte non lié'
        '401':
          description: 'Non authentifié'
        '500':
          description: 'Erreur serveur'
...
  '/api/users/{user_id}':
    get:
      summary: 'Récupérer les détails d''un utilisateur'
      tags:
        - Users
      security:
        -
          bearerAuth: []
      description: 'Récupère les informations complètes d''un utilisateur spécifique'
      parameters:
        -
          in: path
          name: user_id
          required: true
          schema:
            type: string
          description: 'ID de l''utilisateur'
      responses:
        '200':
          description: 'Utilisateur trouvé avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Utilisateur trouvé'}
                  data: {type: object, properties: {user: {type: object, properties: {user_id: {type: string}, email: {type: string}, fullname: {type: string}, avatar_url: {type: string}, role: {type: string, enum: [user, farmer, admin]}, is_verified: {type: boolean}, is_active: {type: boolean}, createdAt: {type: string, format: date-time}, updatedAt: {type: string, format: date-time}, last_login: {type: string, format: date-time}}}}}
        '400':
          description: 'ID utilisateur invalide'
        '401':
          description: 'Non authentifié'
        '404':
          description: 'Utilisateur non trouvé'
        '500':
          description: 'Erreur serveur'
    delete:
      summary: 'Supprimer un utilisateur (soft delete)'
      tags:
        - Users
      security:
        -
          bearerAuth: []
      description: 'Marque un utilisateur comme supprimé (soft delete) - réservée aux administrateurs'
      parameters:
        -
          in: path
          name: user_id
          required: true
          schema:
            type: string
          description: 'ID de l''utilisateur à supprimer'
      responses:
        '200':
          description: 'Utilisateur supprimé avec succès'
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: {type: boolean, example: true}
                  message: {type: string, example: 'Utilisateur supprimé avec succès'}
        '400':
          description: 'ID utilisateur invalide'
        '401':
          description: 'Non authentifié'
        '403':
          description: 'Droits administrateur requis'
        '404':
          description: 'Utilisateur non trouvé'
        '500':
          description: 'Erreur serveur'
...